+++ pk/hpke/README.md
@@ -1,5 +1,7 @@
 # HPKE Example with all supported options
 
+To build wolfSSL for this example run `./configure --enable-hpke --enable-aesgcm --enable-curve25519 --enable-ecc && make && sudo make install`
+
 ```sh
 make
 ./hpke_test
